requirements.txt
# Bibliotecas principales de Computer Vision
opencv-python-headless>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0

# Framework web
streamlit>=1.28.0

# Machine Learning y Data Science
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Bibliotecas Python estándar (incluidas por defecto, listadas para referencia)
# os - módulo estándar
# sys - módulo estándar
# io - módulo estándar
# pickle - módulo estándar
# pathlib - módulo estándar
# importlib - módulo estándar
# subprocess - módulo estándar
# collections - módulo estándar